requirements.txt
aiogram>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
